requirements.txt
Flask==3.0.3
Pillow==10.4.0
numpy==1.26.4

